Fixed permissions in redirects

+++ website/views.py
@@ -1,9 +1,7 @@
-import random
 from database.models import users, mobiles, ipads, cart, MacBooks
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from database.forms import deviceForm
-# csrf_exempt
 from django.views.decorators.csrf import csrf_exempt
 
 # Create your views here.
@@ -33,19 +31,30 @@ def login(request):
             return render(request, "login.html", { 'error': 'Invalid username or password' })
     else:
         username = request.session.get('username')
+        try:
+            synx = int(request.GET.get('synx'))
+        except:
+            synx = 0
         if username:
             return redirect('/')
         else:
-            return render(request, "login.html")
+            if synx == 1:
+                return render(request, "login.html", { 'error': 'Logged out successfully' })
+            elif synx == 2:
+                return render(request, "login.html", { 'error': 'Please login to continue' })
+            elif synx == 3:
+                return render(request, "login.html", { 'error': 'Account created successfully! Please login to continue' })
+            else:
+                return render(request, "login.html")
 
 def logout(request):
     try:
         del request.session['username']
     except:
         pass
-    return redirect('/login')
+    return redirect('/login?synx=1')
 
-def signup(request):
+async def signup(request):
     if request.method == "POST":
         username = request.POST.get('name')
         password = request.POST.get('pswd')
@@ -57,8 +66,7 @@ def signup(request):
             c = cart(name=username, items="")
             user.save()
             c.save()
-            request.session['username'] = username
-            return redirect('/')
+            return redirect('/login?synx=3')
     else:
         username = request.session.get('username')
         if username:
@@ -80,7 +88,10 @@ def menu(request):
         items = MacBooks.objects.all()
         data = list(items)
     else:
-        return redirect('/')
+        # Show 404 page
+        return HttpResponse(
+            '<h1>404 Not Found</h1><br><a href="/">Go back to home page</a>'
+        )
     if username:
         c = cart.objects.get(name=username)
         cart_items = c.items
@@ -102,38 +113,43 @@ def checkout(request):
         return HttpResponse(
             '<h1>Thank you for shopping with us!</h1><br><a href="/">Go back to home page</a>'
         )
-    if username:
-        c = cart.objects.get(name=username)
-        cart_items = c.items
-        cc = []
-        if cart_items:
-            cart_items = cart_items.split(',')
-            cart_total = len(cart_items)
-            for i in cart_items:
-                try:
-                    id = i.split('-')[0]
-                    type = i.split('-')[1]
-                    if type == 'iphones':
-                        item = mobiles.objects.get(id=id)
-                    elif type == 'ipads':
-                        item = ipads.objects.get(id=id)
-                    elif type == 'macbooks':
-                        item = MacBooks.objects.get(id=id)
-                    cc.append(item)
-                except:
-                    pass
-        else:
-            cart_items = 0
-            cc = None
-            cart_total = cart_items
-        print(cc)
-        return render(request, "checkout.html", {'username': username, 'cart_total': cart_total, 'cart_items': cc})
     else:
-        return redirect('/login')
+        if username:
+            c = cart.objects.get(name=username)
+            cart_items = c.items
+            cc = []
+            if cart_items:
+                cart_items = cart_items.split(',')
+                cart_total = len(cart_items)
+                for i in cart_items:
+                    try:
+                        id = i.split('-')[0]
+                        type = i.split('-')[1]
+                        if type == 'iphones':
+                            item = mobiles.objects.get(id=id)
+                        elif type == 'ipads':
+                            item = ipads.objects.get(id=id)
+                        elif type == 'macbooks':
+                            item = MacBooks.objects.get(id=id)
+                        cc.append(item)
+                    except:
+                        pass
+            else:
+                cart_items = 0
+                cc = None
+                cart_total = cart_items
+            print(cc)
+            return render(request, "checkout.html", {'username': username, 'cart_total': cart_total, 'cart_items': cc})
+        else:
+            return redirect('/login?synx=2')
 
 @csrf_exempt
 def product(request):
     if request.method == "POST":
+        username = request.session.get('username')
+        if not username:
+            # redirect to login page with error
+            return redirect('/login?synx=2')
         id = request.POST.get('id')
         type = request.POST.get('type')
         if type == 'iphones':
@@ -163,7 +179,10 @@ def product(request):
         elif type == 'macbooks':
             item = MacBooks.objects.get(id=id)
         else:
-            return redirect('/')
+            # redirect to 404 page
+            return HttpResponse(
+                '<h1>404 Not Found</h1><br><a href="/">Go back to home page</a>'
+            )
         username = request.session.get('username')
         if username:
             c = cart.objects.get(name=username)
@@ -180,15 +199,35 @@ def product(request):
 def manage(request):
     username = request.session.get('username')
     if username != 'admin':
-        return redirect('/')
+        # Show 403 page
+        return HttpResponse(
+            '<h1>403 Forbidden</h1><br><a href="/">Go back to home page</a>'
+        )
     else:
         if request.method == "GET":
             type = request.GET.get('type')
+            try:
+                synx = int(request.GET.get('synx'))
+            except:
+                synx = 0
             if type == "Add":
                 form = deviceForm()
             else:
                 form = None
-            return render(request, "manage.html", {'action': type, 'username': username, 'form': form})
+            if synx == 1:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type, 'error': 'Invalid device type'})
+            elif synx == 2:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type, 'error': 'Device deleted successfully'})
+            elif synx == 3:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type, 'error': 'Invalid form data'})
+            elif synx == 4:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type, 'error': 'Device not found'})
+            elif synx == 5:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type, 'error': 'Device added successfully'})
+            elif synx == 6:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type, 'error': 'Invalid action'})
+            else:
+                return render(request, "manage.html", {'username': username, 'form': form, 'type': type})
         else:
             type = request.POST.get('type')
             action = request.POST.get('action')
@@ -198,36 +237,37 @@ def manage(request):
                     if type == 'iphones':
                             item = mobiles(name=form.cleaned_data['name'], price=form.cleaned_data['price'], image=form.cleaned_data['image'], description=form.cleaned_data['description'])
                             item.save()
-                            return render(request, "manage.html", {'type': type, 'username': username, 'form': form, 'error': 'Item added successfully'})
+                            return redirect('/manage?type=' + action + "&synx=5")
                     elif type == 'ipads':
                             item = ipads(name=form.cleaned_data['name'], price=form.cleaned_data['price'], image=form.cleaned_data['image'], description=form.cleaned_data['description'])
                             item.save()
-                            return render(request, "manage.html", {'type': type, 'username': username, 'form': form, 'error': 'Item added successfully'})
+                            return redirect('/manage?type=' + action + "&synx=5")
                     elif type == 'macbooks':
                             item = MacBooks(name=form.cleaned_data['name'], price=form.cleaned_data['price'], image=form.cleaned_data['image'], description=form.cleaned_data['description'])
                             item.save()
-                            return render(request, "manage.html", {'type': type, 'username': username, 'form': form, 'error': 'Item added successfully'})
+                            return redirect('/manage?type=' + action + "&synx=5")
                     else:
-                        return render(request, "manage.html", {'type': type, 'username': username, 'form': form, 'error': 'Invalid type'})
+                        return redirect('/manage?type=' + action + "&synx=1")
                 else:
-                    print(form.errors)
-                    return redirect('/manage?type=' + action)
-            else:
+                    return redirect('/manage?type=' + action + "&synx=3")
+            elif action == "Remove":
                 id = request.POST.get('id')
                 try:
                     if type == 'iphones':
                         item = mobiles.objects.get(id=id)
                         item.delete()
-                        return render(request, "manage.html", {'type': type, 'username': username, 'error': 'Item deleted successfully'})
+                        return redirect('/manage?type=' + action + "&synx=2")
                     elif type == 'ipads':
                         item = ipads.objects.get(id=id)
                         item.delete()
-                        return render(request, "manage.html", {'type': type, 'username': username, 'error': 'Item deleted successfully'})
+                        return redirect('/manage?type=' + action + "&synx=2")
                     elif type == 'macbooks':
                         item = MacBooks.objects.get(id=id)
                         item.delete()
-                        return render(request, "manage.html", {'type': type, 'username': username, 'error': 'Item deleted successfully'})
+                        return redirect('/manage?type=' + action + "&synx=2")
                     else:
-                        return render(request, "manage.html", {'type': type, 'username': username, 'error': 'Invalid type'})
+                        return redirect('/manage?type=' + action + "&synx=1")
                 except:
-                    return render(request, "manage.html", {'type': type, 'username': username, 'error': 'Item not found'})
+                    return redirect('/manage?type=' + action + "&synx=4")
+            else:
+                return redirect('/manage?type=' + action + "&synx=6")
